+++ bin/actions/upload_archive_action.js
@@ -1,0 +1,12 @@
+const StatusLog = require('../support/status_log')
+
+const uploadArchiveAction = (resourceBundle, manifestConfiguration) => {
+  const { archiveBundle, TransitTool } = resourceBundle
+
+  StatusLog.notify(`upload ${archiveBundle.uri}`, archiveBundle.uuid)
+  return TransitTool
+    .push(archiveBundle, manifestConfiguration)
+    .then(() => resourceBundle)
+}
+
+module.exports = uploadArchiveAction