+++ bin/packages/folder_package.js
@@ -1,7 +1,21 @@
+const FileSystem = require('../support/file_system')
 const fs = require('fs-extra')
 
 class FolderPackage {
-  static sendToStaging ({ cachePath, stagingPath }) {
+  static pack (archiveBundle, manifestConfiguration) {
+    return new Promise((resolve, reject) => {
+      FileSystem.copyNonIgnoredFiles(
+        archiveBundle.releaseFolder,
+        archiveBundle.releaseStaging,
+        manifestConfiguration.ignore()
+      ).then(() => {
+        return {
+          releaseAsset: archiveBundle.releaseStaging
+        }
+      })
+    })
+  }
+  static unpack ({ cachePath, stagingPath }) {
     return new Promise((resolve, reject) => {
       fs.copy(cachePath, stagingPath, err => {
         if (err) {