+++ bin/packages/tar_package.js
@@ -2,18 +2,33 @@ const FileSystem = require('../support/file_system')
 const tar = require('tar')
 
 class TarPackage {
-  static sendToStaging ({ cachePath, stagingPath }) {
-    return new Promise((resolve, reject) => {
-      tar.extract({
-        cwd: stagingPath,
-        file: FileSystem.readPath(cachePath),
-        preserveOwner: false,
-        unlink: true,
-        strip: 1
+  static pack (archiveBundle, manifestConfiguration) {
+    const releaseAsset = `${archiveBundle.releaseStaging}.tar`
+
+    return tar.create({
+      gzip: false,
+      file: FileSystem.readPath(releaseAsset)
+    }, [archiveBundle.releaseFolder])
+      .then(() => {
+        return {
+          releaseAsset
+        }
       })
-      .then(() => resolve({}))
-      .catch((reason) => reject(reason))
+  }
+  static unpack ({
+    cachePath,
+    stagingPath
+  }) {
+    return tar.extract({
+      cwd: stagingPath,
+      file: FileSystem.readPath(cachePath),
+      preserveOwner: false,
+      unlink: true,
+      strip: 1
     })
+      .then(() => {
+        return {}
+      })
   }
 }