+++ bin/support/status_log.js
@@ -1,5 +1,3 @@
-const spinners = require('cli-spinners')
-const readline = require('readline')
 const winston = require('winston')
 const color = require('cli-color')
 const process = require('process')
@@ -11,17 +9,7 @@ const Colors = {
   red: color.redBright
 }
 
-const Spinner = {
-  running: spinners.hamburger.frames
-}
-
 class StatusLog {
-  static __drawLine (content) {
-    if (this.__stream) {
-      this.__stream.write(null, { ctrl: true, name: 'u' })
-      this.__stream.write(content)
-    }
-  }
 
   static __drawContent () {
     const spinner = Spinner.running[this.__frame % Spinner.running.length]
@@ -29,41 +17,25 @@ class StatusLog {
 
     const errorCount = _.size(this.__errors)
     const errors = (errorCount > 0) ? Colors.red(errorCount + '! ') : ``
+    const action = ``
 
-    return `${Colors.yellow(spinner)} ${Colors.gray(seconds + 's')} ${errors}${this.__action}`
+    return `${Colors.yellow(spinner)} ${Colors.gray(seconds + 's')} ${errors}${action}`
   }
 
   static start () {
-    if (!_.includes(process.argv, `--verbose`)) {
-      this.__interval = setInterval(() => {
-        this.__frame += 1
-        this.__drawLine(this.__drawContent())
-      }, this.__refreshRate)
-    }
-
     return this
   }
 
   static stop () {
-    clearInterval(this.__interval)
-    if (this.__stream) {
-      this.__stream.write(null, { ctrl: true, name: 'u' })
-    }
-
     return this
   }
 
-  static initialize () {
+  static initialize (vorpal) {
     this.uninitialize()
 
+    this.__vorpal = vorpal
+
     const transports = [
-      new winston.transports.Console({
-        level: 'error',
-        json: false,
-        formatter: (info) => {
-          return (info.message) ? `${Colors.red('Failure(s)')}\n${info.message}` : ``
-        }
-      }),
       new (winston.transports.File)({
         timestamp: () => Date.now(),
         handleExceptions: true,
@@ -75,22 +47,10 @@ class StatusLog {
       })
     ]
 
-    if (_.includes(process.argv, `--verbose`)) {
-      transports.push(new winston.transports.Console({
-
-      }))
-    }
-
     this.__logger = new (winston.Logger)({
       exitOnError: false,
       level: 'info',
       transports
-    })
-
-    this.__stream = readline.createInterface({
-      output: process.stdout,
-      input: process.stdin,
-      prompt: ``
     })
 
     return this
@@ -98,17 +58,17 @@ class StatusLog {
 
   static uninitialize () {
     this.__logger = undefined
-    this.__stream = undefined
-    this.__refreshRate = 128
-    this.__frame = 0
+    this.__vorpal = undefined
     this.__errors = []
-    this.__action = ''
+    this.__frame = 0
+
     return this
   }
 
   static notify (action, resource, meta = {}) {
     this.__logger.info(`[${resource}] ${action}`, meta)
-    this.__action = action
+    this.__vorpal.log(action)
+
     return this
   }
 
@@ -118,27 +78,23 @@ class StatusLog {
   }
 
   static completeSuccess () {
-    this.__action = 'completed'
+    this.notify(`completed`)
 
     return new Promise((resolve, reject) => {
-      setTimeout(() => {
-        resolve(this.stop().__writeErrors().uninitialize())
-      }, 1000)
+      resolve(this.stop().__writeErrors().uninitialize())
     })
   }
 
   static completeFailure (reason) {
-    this.__action = 'failed'
+    this.notify(`failed`)
 
     StatusLog.error(reason)
     return new Promise((resolve, reject) => {
-      setTimeout(() => {
-        resolve(this.stop().__writeErrors().uninitialize())
-      }, 1000)
+      resolve(this.stop().__writeErrors().uninitialize())
     })
   }
 
-  static __writeErrors() {
+  static __writeErrors () {
     const errors = _.reduce(this.__errors, (content, reason, index) => {
       return `${content}${index + 1}. ${Colors.gray(reason)}\n`
     }, ``)